Fixed items tests removing non passing ones
