+++ .env.sample
@@ -1,0 +1,2 @@
+PORT=8080
+MONGO=mongodb://user:pass@host:port/dbname
