+++ package.json
@@ -30,7 +30,7 @@
     "events": "^1.0.2",
     "express": "^4.11.2",
     "express-session": "^1.10.2",
-    "mongoose": "^3.8.23",
+    "mongoose": "^3.8.25",
     "passport": "^0.2.1",
     "passport-local": "^1.0.0",
     "restify": "^2.8.5",
